requirements.txt
fastapi
uvicorn[standard]
razorpay
python-dotenv
pydantic
httpx
supabase
